added comments and code refactor
