+++ app/src/main/java/com/likeminds/feedsx/posttypes/view/adapter/databinder/postdocuments/ItemDocumentViewDataBinder.kt
@@ -21,7 +21,6 @@ class ItemDocumentViewDataBinder : ViewDataBinder<ItemDocumentBinding, Attachmen
         data: AttachmentViewData,
         position: Int
     ) {
-
         // sets the document data on the doc view
         PostTypeUtil.initDocument(
             binding,
